# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing device tokens and delivering push notifications via FCM, APNS, and Web Push.",
        "title": "Push Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/devices/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Registers a new device token or updates an existing one for push notifications.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Register or update a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device details",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_onix-fun_push-service_internal_model.Device"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid JSON or missing fields",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Marks a device as inactive so it no longer receives notifications.",
                "tags": [
                    "devices"
                ],
                "summary": "Deactivate a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/recipients/{recipient}/devices": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all devices registered for a specific recipient ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipients"
                ],
                "summary": "List devices for a recipient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipient ID",
                        "name": "recipient",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_onix-fun_push-service_internal_model.Device"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_onix-fun_push-service_internal_model.Device": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "recipient_id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}